updated 25 august 2022

+++ app/Http/Controllers/Admin/PackageController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Package;
 use App\PackageContent;
 use App\PackageDetail;
+use App\Subscription;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -15,6 +16,7 @@ class PackageController extends Controller
     public function index()
     {
         $data['packages'] = Package::get();
+
         return view('admin.package.list', $data);
     }
 
@@ -23,7 +25,8 @@ class PackageController extends Controller
     public function add()
     {
         $data['pack_desc'] = PackageContent::get();
-        return view('admin.package.add',$data);
+
+        return view('admin.package.add', $data);
     }
 
     //edit page of packages
@@ -31,9 +34,22 @@ class PackageController extends Controller
     public function edit($id)
     {
         // $data['pack_desc'] = PackageContent::get();
-        $data['pack_desc']=DB::select('SELECT * FROM package_contents  WHERE package_contents.id NOT IN (SELECT package_details.package_content_id FROM package_details WHERE package_details.package_id='.$id.' )');
+        $data['pack_desc'] = DB::select('SELECT * FROM package_contents  WHERE package_contents.id NOT IN (SELECT package_details.package_content_id FROM package_details WHERE package_details.package_id='.$id.' )');
         $data['package'] = Package::find($id);
-        $data['package_detail'] = PackageDetail::with('package_content')->where('package_id',$id)->get();
+        $check = PackageDetail::with('package_content')->where('package_id', $id)->get();
+        if (count($check) == 0) {
+            $package_content = PackageContent::get();
+            foreach ($package_content as $key => $value) {
+                // code...
+                $package_detail = new PackageDetail();
+                $package_detail->package_id = $id;
+                $package_detail->package_content_id = $value->id;
+                $package_detail->status = 0;
+                $package_detail->save();
+            }
+        }
+        $data['package_detail'] = PackageDetail::with('package_content')->where('package_id', $id)->get();
+
         return view('admin.package.edit', $data);
     }
 
@@ -41,24 +57,26 @@ class PackageController extends Controller
 
     public function store(Request $request)
     {
-        $package = new Package;
+        $package = new Package();
         // dd($request);
         $package->name = $request->name;
+        $package->sound_tracks = $request->sound_track;
+        $package->sound_effects = $request->sound_effect;
         $package->downloads = $request->downloads;
+        $package->content = $request->content;
         $package->price = $request->price;
 
         $package->save();
 
-        if(count($request->package_detail_id) != 0 ){
-            for ($i=0; $i < count($request->package_detail_id) ; $i++) {
-                # code...
-                $package_detail=new PackageDetail;
-                $package_detail->package_id=$package->id;
-                $package_detail->package_content_id=$request->package_detail_id[$i];
-                $package_detail->status=$request->status[$i];
+        if (count($request->package_detail_id) != 0) {
+            for ($i = 0; $i < count($request->package_detail_id); ++$i) {
+                // code...
+                $package_detail = new PackageDetail();
+                $package_detail->package_id = $package->id;
+                $package_detail->package_content_id = $request->package_detail_id[$i];
+                $package_detail->status = $request->status[$i];
                 $package_detail->save();
             }
-
         }
 
         return redirect('/admin/package')->with('success', 'Package is successfully saved');
@@ -67,25 +85,27 @@ class PackageController extends Controller
     //update package
     public function update($id, Request $request)
     {
-        $package =  Package::find($id);;
+        $package = Package::find($id);
         // dd($request);
         $package->name = $request->name;
         $package->downloads = $request->downloads;
+        $package->sound_tracks = $request->sound_track;
+        $package->sound_effects = $request->sound_effect;
+        $package->content = $request->content;
         $package->price = $request->price;
         $package->save();
 
-        if(count($request->package_detail_id) != 0 ){
+        if (count($request->package_detail_id) != 0) {
             DB::table('package_details')->where('package_id', $id)->delete();
 
-            for ($i=0; $i < count($request->package_detail_id) ; $i++) {
-                # code...
-                $package_detail=new PackageDetail;
-                $package_detail->package_id=$package->id;
-                $package_detail->package_content_id=$request->package_detail_id[$i];
-                $package_detail->status=$request->status[$i];
+            for ($i = 0; $i < count($request->package_detail_id); ++$i) {
+                // code...
+                $package_detail = new PackageDetail();
+                $package_detail->package_id = $package->id;
+                $package_detail->package_content_id = $request->package_detail_id[$i];
+                $package_detail->status = $request->status[$i];
                 $package_detail->save();
             }
-
         }
 
         return redirect('/admin/package')->with('success', 'Package is successfully Updated');
@@ -94,36 +114,52 @@ class PackageController extends Controller
     //delete package
     public function destroy($id)
     {
-        $package = Package::where('id', $id);
+        $check = Subscription::where('package_id', $id)->count();
+        if ($check == 0 && $id != 1) {
+            $package = Package::where('id', $id);
+            $packagedec = PackageDetail::where('package_id', $id);
 
-        $package->delete();
-        return back();
+            $packagedec->delete();
+            $package->delete();
+
+            return back();
+        }
+
+        return back()->with('error', 'Package can`t be deleted');
     }
-
 
     //  Package Description
-    public function packageDesc(){
+    public function packageDesc()
+    {
         $data['pack_desc'] = PackageContent::get();
-        return view('admin.package.package_details',$data);
+
+        return view('admin.package.package_details', $data);
     }
 
-    public function packageDescStore(Request $request){
-        $pack = new PackageContent;
+    public function packageDescStore(Request $request)
+    {
+        $pack = new PackageContent();
         $pack->description = $request->content;
         $pack->save();
+
         return redirect('admin/pack/description')->with('success', 'Package Description is successfully saved');
     }
-    public function packageDescUpdate(Request $request){
+
+    public function packageDescUpdate(Request $request)
+    {
         $pack = PackageContent::find($request->id);
         $pack->description = $request->content;
         $pack->save();
-        return redirect('admin/pack/description')->with('success', 'Package Description is successfully Updated');
 
+        return redirect('admin/pack/description')->with('success', 'Package Description is successfully Updated');
     }
-    public function packageDescdestroy($id){
+
+    public function packageDescdestroy($id)
+    {
         $package = PackageContent::where('id', $id);
 
         $package->delete();
+
         return back();
     }
 }
